feat: task implementation

+++ practice2025/CommandRunner/CommandRunner.cs
@@ -2,7 +2,7 @@
 
 using System.Reflection;
 
-class Program
+public class CommandRunner
 {
     public static void Main()
     {
@@ -10,9 +10,8 @@ class Program
 
         if (baseDir is null)
             throw new DirectoryNotFoundException("Couldn't find base directory!\n");
-        
 
-        string? DllPath = Directory.GetFiles(baseDir, "FileSystemCommands.dll", SearchOption.AllDirectories).First();
+        string? DllPath = Directory.GetFiles(baseDir, "FileSystemCommands.dll", SearchOption.AllDirectories).FirstOrDefault();
 
         if (DllPath is null)
             throw new FileNotFoundException("Couldn't find dynamic library!\n");
@@ -31,14 +30,25 @@ class Program
         var assembly = Assembly.LoadFrom(DllPath);
 
         var directorySizeCommand = assembly.GetType("FileSystemCommands.DirectorySizeCommand");
+        if (directorySizeCommand is null)
+            throw new TypeLoadException("Couldn't load type 'FileSystemCommands.DirectorySizeCommand'.");
+
         var directorySizeCommandInstance = Activator.CreateInstance(directorySizeCommand, new object[] { testDir });
-        directorySizeCommandInstance.GetType().GetMethod("Execute")?.Invoke(directorySizeCommandInstance, null);
+        directorySizeCommandInstance?.GetType().GetMethod("Execute")?.Invoke(directorySizeCommandInstance, null);
+        var size = directorySizeCommandInstance?.GetType().GetProperty("Size")?.GetValue(directorySizeCommandInstance, null);
+
+        Console.WriteLine(size);
 
         var findFilesCommand = assembly.GetType("FileSystemCommands.FindFilesCommand");
+        if (findFilesCommand is null)
+            throw new TypeLoadException("Couldn't load type 'FileSystemCommands.FindFilesCommand'.");
+
         var findFilesCommandInstance = Activator.CreateInstance(findFilesCommand, new object[] { testDir, mask });
-        findFilesCommandInstance.GetType().GetMethod("Execute")?.Invoke(findFilesCommandInstance, null);
+        findFilesCommandInstance?.GetType().GetMethod("Execute")?.Invoke(findFilesCommandInstance, null);
+        var files = findFilesCommandInstance?.GetType().GetProperty("Files")?.GetValue(findFilesCommandInstance, null) as List<string>;
+
+        files?.ForEach(file => Console.WriteLine(file));
 
         Directory.Delete(testDir, true);
-
     }
 }
